DAY-6 refactor

+++ src/main/scala/day06/model/RaceData.scala
@@ -1,3 +1,3 @@
 package day06.model
 
-case class RaceData(length: Int, recordDistance: Int)
+case class RaceData(length: Long, recordDistance: Long)
